+++ app/src/modules/files/routes/item.vue
@@ -2,7 +2,7 @@
 	<files-not-found v-if="!loading && !item" />
 	<private-view v-else :title="loading || !item ? $t('loading') : item.title">
 		<template #title-outer:prepend>
-			<v-button class="header-icon" rounded icon secondary exact :to="to">
+			<v-button class="header-icon" rounded icon secondary exact @click="$router.back()">
 				<v-icon name="arrow_back" />
 			</v-button>
 		</template>
@@ -99,7 +99,7 @@
 
 				<template #append-outer>
 					<save-options
-						:disabled="hasEdits === false || saveAllowed === false"
+						v-if="hasEdits === true || saveAllowed === true"
 						@save-and-stay="saveAndStay"
 						@save-as-copy="saveAsCopyAndNavigate"
 					/>
@@ -137,6 +137,7 @@
 				:batch-mode="isBatch"
 				:primary-key="primaryKey"
 				:disabled="updateAllowed === false"
+				:validation-errors="validationErrors"
 				v-model="edits"
 			/>
 		</div>
@@ -157,7 +158,7 @@
 		<template #sidebar>
 			<file-info-sidebar-detail :file="item" />
 			<revisions-drawer-detail
-				v-if="isBatch === false && isNew === false"
+				v-if="isBatch === false && isNew === false && revisionsAllowed"
 				collection="directus_files"
 				:primary-key="primaryKey"
 				ref="revisionsDrawerDetail"
@@ -176,33 +177,26 @@
 <script lang="ts">
 import { defineComponent, computed, toRefs, ref, watch } from '@vue/composition-api';
 import FilesNavigation from '../components/navigation.vue';
-import { i18n } from '../../../lang';
-import router from '../../../router';
-import RevisionsDrawerDetail from '../../../views/private/components/revisions-drawer-detail';
-import CommentsSidebarDetail from '../../../views/private/components/comments-sidebar-detail';
-import useItem from '../../../composables/use-item';
-import SaveOptions from '../../../views/private/components/save-options';
-import FilePreview from '../../../views/private/components/file-preview';
-import ImageEditor from '../../../views/private/components/image-editor';
-import { nanoid } from 'nanoid';
-import FileLightbox from '../../../views/private/components/file-lightbox';
-import { useFieldsStore } from '../../../stores/';
-import { Field } from '../../../types';
+import { i18n } from '@/lang';
+import router from '@/router';
+import RevisionsDrawerDetail from '@/views/private/components/revisions-drawer-detail';
+import CommentsSidebarDetail from '@/views/private/components/comments-sidebar-detail';
+import useItem from '@/composables/use-item';
+import SaveOptions from '@/views/private/components/save-options';
+import FilePreview from '@/views/private/components/file-preview';
+import ImageEditor from '@/views/private/components/image-editor';
+import { Field } from '@/types';
 import FileInfoSidebarDetail from '../components/file-info-sidebar-detail.vue';
-import useFormFields from '../../../composables/use-form-fields';
 import FolderPicker from '../components/folder-picker.vue';
-import api, { addTokenToURL } from '../../../api';
-import getRootPath from '../../../utils/get-root-path';
+import api, { addTokenToURL } from '@/api';
+import { getRootPath } from '@/utils/get-root-path';
 import FilesNotFound from './not-found.vue';
-import useShortcut from '../../../composables/use-shortcut';
+import useShortcut from '@/composables/use-shortcut';
 import ReplaceFile from '../components/replace-file.vue';
-import { usePermissions } from '../../../composables/use-permissions';
-import { notify } from '../../../utils/notify';
-import { unexpectedError } from '../../../utils/unexpected-error';
-
-type Values = {
-	[field: string]: any;
-};
+import { usePermissions } from '@/composables/use-permissions';
+import { notify } from '@/utils/notify';
+import { unexpectedError } from '@/utils/unexpected-error';
+import unsavedChanges from '@/composables/unsaved-changes';
 
 export default defineComponent({
 	name: 'files-item',
@@ -240,7 +234,6 @@ export default defineComponent({
 		const form = ref<HTMLElement>();
 		const { primaryKey } = toRefs(props);
 		const { breadcrumb } = useBreadcrumb();
-		const fieldsStore = useFieldsStore();
 		const replaceFileDialogActive = ref(false);
 
 		const revisionsDrawerDetail = ref<Vue | null>(null);
@@ -258,16 +251,19 @@ export default defineComponent({
 			saveAsCopy,
 			isBatch,
 			refresh,
+			validationErrors,
 		} = useItem(ref('directus_files'), primaryKey);
 
 		const hasEdits = computed<boolean>(() => Object.keys(edits.value).length > 0);
+
+		unsavedChanges(hasEdits);
+
 		const confirmDelete = ref(false);
 		const editActive = ref(false);
 		const fileSrc = computed(() => {
 			if (item.value && item.value.modified_on) {
 				return addTokenToURL(
-					getRootPath() +
-						`assets/${props.primaryKey}?cache-buster=${item.value.modified_on}&key=system-large-contain`
+					getRootPath() + `assets/${props.primaryKey}?cache-buster=${item.value.modified_on}&key=system-large-contain`
 				);
 			}
 
@@ -303,7 +299,7 @@ export default defineComponent({
 
 		useShortcut('meta+s', saveAndStay, form);
 
-		const { deleteAllowed, saveAllowed, updateAllowed, fields } = usePermissions(
+		const { deleteAllowed, saveAllowed, updateAllowed, fields, revisionsAllowed } = usePermissions(
 			ref('directus_files'),
 			item,
 			isNew
@@ -349,6 +345,8 @@ export default defineComponent({
 			updateAllowed,
 			fields,
 			fieldsFiltered,
+			revisionsAllowed,
+			validationErrors,
 		};
 
 		function useBreadcrumb() {
@@ -381,7 +379,7 @@ export default defineComponent({
 		async function saveAndQuit() {
 			try {
 				await save();
-				router.push(`/files`);
+				router.push(to.value);
 			} catch {
 				// `save` will show unexpected error dialog
 			}
@@ -398,13 +396,13 @@ export default defineComponent({
 
 		async function saveAsCopyAndNavigate() {
 			const newPrimaryKey = await saveAsCopy();
-			router.push(`/files/${newPrimaryKey}`);
+			if (newPrimaryKey) router.push(`/files/${newPrimaryKey}`);
 		}
 
 		async function deleteAndQuit() {
 			try {
 				await remove();
-				router.push(`/files`);
+				router.push(to.value);
 			} catch {
 				// `remove` will show the unexpected error dialog
 			}
@@ -470,9 +468,9 @@ export default defineComponent({
 
 <style lang="scss" scoped>
 .action-delete {
-	--v-button-background-color: var(--danger-25);
+	--v-button-background-color: var(--danger-10);
 	--v-button-color: var(--danger);
-	--v-button-background-color-hover: var(--danger-50);
+	--v-button-background-color-hover: var(--danger-25);
 	--v-button-color-hover: var(--danger);
 }
 
@@ -483,9 +481,9 @@ export default defineComponent({
 .edit,
 .folder,
 .download {
-	--v-button-background-color: var(--primary-25);
+	--v-button-background-color: var(--primary-10);
 	--v-button-color: var(--primary);
-	--v-button-background-color-hover: var(--primary-50);
+	--v-button-background-color-hover: var(--primary-25);
 	--v-button-color-hover: var(--primary);
 }
